30-12-21 partial changes

+++ app/Http/Controllers/JournalVoucherController.php
@@ -124,20 +124,23 @@ class JournalVoucherController extends Controller
         $check = false;
         $remainingBalance = 0;
         $remainingBalanceType = '';
-        if ($action && str_replace(',','',$request->suspense_amount) > 0 && (array_sum($request->debit_amounts) > array_sum($request->credit_amounts) || array_sum($request->credit_amounts) > array_sum($request->debit_amounts))) {
-            $suspenseEntryDetail = $voucher->voucherDetails()->where('suspense_account', '1')->first() != null ? $voucher->voucherDetails()->where('suspense_account', '1')->first() : new VoucherDetail();
+        if($action && str_replace(',','',$request->suspense_amount) > 0 && (array_sum(str_replace(',','',array_values($request->debit_amounts)))>array_sum(str_replace(',','',array_values($request->credit_amounts))) || array_sum(str_replace(',','',array_values($request->credit_amounts)))>array_sum(str_replace(',','',array_values($request->debit_amounts))))){
+            $suspenseEntryDetail = $voucher->voucherDetails()->where('suspense_account','1')->first() !=null ?$voucher->voucherDetails()->where('suspense_account','1')->first():new VoucherDetail();
             $check = true;
-        } else if ($action && (array_sum($request->debit_amounts) == array_sum($request->credit_amounts))) {
-            if ($voucher->voucherDetails()->where('suspense_account', '1')->first() != null) {
-                VoucherDetail::where('id', $voucher->voucherDetails()->where('suspense_account', '1')->first()->id)->delete();
+            return "first";
+        }else if($action && str_replace(',','',$request->suspense_amount) > 0 && (array_sum(str_replace(',','',array_values($request->debit_amounts)))==array_sum(str_replace(',','',array_values($request->credit_amounts))))){
+            return "second";
+            if($voucher->voucherDetails()->where('suspense_account','1')->first() !=null){
+                VoucherDetail::where('id',$voucher->voucherDetails()->where('suspense_account','1')->first()->id)->delete();
             }
-        } else if (!$action && str_replace(',','',$request->suspense_amount) > 0 && (array_sum($request->debit_amounts) > array_sum($request->credit_amounts) || array_sum($request->credit_amounts) > array_sum($request->debit_amounts))) {
+        }else if(!$action && str_replace(',','',$request->suspense_amount) > 0 && (array_sum(str_replace(',','',array_values($request->debit_amounts)))>array_sum(str_replace(',','',array_values($request->credit_amounts))) || array_sum(str_replace(',','',array_values($request->credit_amounts)))>array_sum(str_replace(',','',array_values($request->debit_amounts))))){
             $suspenseEntryDetail = new VoucherDetail();
+            return "third";
             $check = true;
         }
-
-        $openingBalance = isset($request->suspense_account)? getOpeningBalance($request->suspense_account,$request->suspense_date,false,0)["opening_balance"]:0;
-        $transactionType = isset($request->suspense_account)? getOpeningBalance($request->suspense_account,$request->suspense_date,false,0)["opening_balance_type"]:'';
+        
+        $openingBalance = isset($request->suspense_account) && isset($request->suspense_date)? getOpeningBalance($request->suspense_account,$request->suspense_date,false,0)["opening_balance"]:0;
+        $transactionType = isset($request->suspense_account) && isset($request->suspense_date)? getOpeningBalance($request->suspense_account,$request->suspense_date,false,0)["opening_balance_type"]:'';
 
         if ($transactionType == "debit" && $request->suspense_entry == "debit") {
             $remainingBalance = $openingBalance + str_replace(',','',$request->suspense_amount);
